requirements.txt
pillow
fastapi
filetype
uvicorn[standard]
python-multipart